+++ test/Spec.hs
@@ -1,8 +1,6 @@
-import qualified AlphaEqTest
-import qualified BlameTest
 import Control.Lens.Operators
-import qualified LangATest
-import qualified LangBTest
+import qualified PlainTest
+import PolyKindsTH ()
 import Test.Tasty
 
 import Prelude
@@ -11,8 +9,5 @@ main :: IO ()
 main =
     testGroup
         "Tests"
-        [ testGroup "infer" [LangATest.test, LangBTest.test]
-        , AlphaEqTest.test
-        , BlameTest.test
-        ]
+        [PlainTest.test]
         & defaultMain